+++ go/api/stream_commands.go
@@ -49,4 +49,109 @@ type StreamCommands interface {
 	//
 	// [valkey.io]: https://valkey.io/commands/xadd/
 	XAddWithOptions(key string, values [][]string, options *options.XAddOptions) (Result[string], error)
+
+	// Changes the ownership of a pending message.
+	//
+	// See [valkey.io] for details.
+	//
+	// Parameters:
+	//  key      - The key of the stream.
+	//  group    - The name of the consumer group.
+	//  consumer - The name of the consumer.
+	//  minIdleTime - The minimum idle time in milliseconds.
+	//  ids        - The ids of the entries to claim.
+	//
+	// Return value:
+	//  A `map of message entries with the format `{"entryId": [["entry", "data"], ...], ...}` that were claimed by
+	//  the consumer.
+	//
+	// Example:
+	//
+	// [valkey.io]: https://valkey.io/commands/xclaim/
+	XClaim(
+		key string,
+		group string,
+		consumer string,
+		minIdleTime int64,
+		ids []string,
+	) (map[Result[string]][][]Result[string], error)
+
+	// Changes the ownership of a pending message.
+	//
+	// See [valkey.io] for details.
+	//
+	// Parameters:
+	//  key      - The key of the stream.
+	//  group    - The name of the consumer group.
+	//  consumer - The name of the consumer.
+	//  minIdleTime - The minimum idle time in milliseconds.
+	//  ids        - The ids of the entries to claim.
+	//  options    - Stream claim options.
+	//
+	// Return value:
+	//  A `map` of message entries with the format `{"entryId": [["entry", "data"], ...], ...}` that were claimed by
+	//  the consumer.
+	//
+	// Example:
+	//  result, err := ...
+	//
+	// [valkey.io]: https://valkey.io/commands/xclaim/
+	XClaimWithOptions(
+		key string,
+		group string,
+		consumer string,
+		minIdleTime int64,
+		ids []string,
+		options *options.StreamClaimOptions,
+	) (map[Result[string]][][]Result[string], error)
+
+	// Changes the ownership of a pending message. This function returns an `array` with
+	// only the message/entry IDs, and is equivalent to using `JUSTID` in the Valkey API.
+	//
+	// See [valkey.io] for details.
+	//
+	// Parameters:
+	//  key      - The key of the stream.
+	//  group    - The name of the consumer group.
+	//  consumer - The name of the consumer.
+	//  minIdleTime - The minimum idle time in milliseconds.
+	//  ids        - The ids of the entries to claim.
+	//  options    - Stream claim options.
+	//
+	// Return value:
+	//  An array of the ids of the entries that were claimed by the consumer.
+	//
+	// Example:
+	//  result, err := ...
+	// [valkey.io]: https://valkey.io/commands/xclaim/
+	XClaimJustId(key string, group string, consumer string, minIdleTime int64, ids []string) ([]Result[string], error)
+
+	// Changes the ownership of a pending message. This function returns an `array` with
+	// only the message/entry IDs, and is equivalent to using `JUSTID` in the Valkey API.
+	//
+	// See [valkey.io] for details.
+	//
+	// Parameters:
+	//  key      - The key of the stream.
+	//  group    - The name of the consumer group.
+	//  consumer - The name of the consumer.
+	//  minIdleTime - The minimum idle time in milliseconds.
+	//  ids        - The ids of the entries to claim.
+	//  options    - Stream claim options.
+	//
+	// Return value:
+	//  An array of the ids of the entries that were claimed by the consumer.
+	//
+	// Example:
+	//  result, err := ...
+	//
+	// [valkey.io]: https://valkey.io/commands/xclaim/
+	XClaimJustIdWithOptions(
+		key string,
+		group string,
+		consumer string,
+		minIdleTime int64,
+		ids []string,
+		options *options.StreamClaimOptions,
+	) ([]Result[string], error)
 }