+++ go/integTest/shared_commands_test.go
@@ -4456,3 +4456,12 @@ func (suite *GlideTestSuite) TestZRem() {
 		assert.IsType(suite.T(), &api.RequestError{}, err)
 	})
 }
+
+func (suite *GlideTestSuite) TestXClaim() {
+	suite.runWithDefaultClients(func(client api.BaseClient) {
+		key := uuid.New().String()
+		client.XAdd(key, [][]string{{"field1", "value1"}})
+
+		// TODO: finish test case using custom commands
+	})
+}